python-backend/requirements.txt
# requirements.txt
heapq
fastapi
uvicorn[standard]
motor
pydantic
scikit-learn  # For ML in analyzers
tensorflow   # For advanced ML if needed
paho-mqtt